+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps/scripts/modules/data/BoundingBox.ts
@@ -1,0 +1,8 @@
+import * as atlas from "azure-maps-control"
+
+export class BoundingBox {
+    public static fromData(data: any) {
+        const bbox = atlas.data.BoundingBox.fromData(data);
+        return bbox;
+    }
+}