+++ Documents/ReleaseNotes.md
@@ -2,7 +2,10 @@
 
 ### [<- Go Back](../README.md)
 - `11.0.0-Preview-3.0`:
-  - `IAzureMapsInterop.Math`. New interop module based on the `azure-maps-control` Math class.
+ - `IAzureMapsInterop.BoundingBox`. New interop module based on the azure maps
+   [BoundingBox Class](https://learn.microsoft.com/en-us/javascript/api/azure-maps-control/atlas.data.boundingbox?view=azure-maps-typescript-latest).
+  - `IAzureMapsInterop.Math`. New interop module based on the azure maps
+   [Math Class](https://learn.microsoft.com/en-us/javascript/api/azure-maps-control/atlas.math?view=azure-maps-typescript-latest).
   - `IAzureMapsSources`
 	- `GetJSShapeById`. New method to get an IJSObjectReference for a shape in the source.
 	- `GetJSShapesById`. New method to get a list of IJSObjectReferences for the shapes in a source by shape id.